Implemented Automated Test Code
All motors with encoders can now be tested individually or in sequence.
The only requirement on the sequence is that you should manually verify
the Solenoid and change Test Safety to true before it will run. The
tests check that when a positive joystick value is given to the motor
the encoder goes up. We need to verify this.